added CarImages table

+++ Entites/Concrete/Car.cs
@@ -14,7 +14,7 @@ namespace Entites.Concrete
         public int BrandId { get; set; }
         public int ColorId { get; set; }
         public int ModelYear { get; set; }
-        public short DailyPrice { get; set; }
+        public int DailyPrice { get; set; }
         public string? Description { get; set; }
     }
 }
